refactor: refatorando implementação do database
