+++ projects/backend/package.json
@@ -18,8 +18,8 @@
     "date-fns": "^2.30.0",
     "env-var": "^7.3.1",
     "jsonwebtoken": "^9.0.0",
-    "mysql2": "^3.3.3",
     "node-schedule": "^2.1.1",
+    "pg": "^8.11.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1",
     "typeorm": "^0.3.16",
