+++ projects/backend/src/infra/database/index.ts
@@ -1,3 +1,3 @@
 export type { Database } from './database'
 export { LessThanOrEqual } from './operators'
-export { makeDatabase } from './make-database'
+export { makePostgresDatabase } from './typeorm'
