--- projects/backend/src/infra/database/mysql/index.ts
@@ -1,1 +1,0 @@
-export { MysqlDatabase } from './mysql.database'
