+++ projects/backend/src/infra/database/typeorm/index.ts
@@ -1,0 +1,1 @@
+export { makePostgresDatabase } from './make-typeorm-database'
