+++ projects/backend/src/infra/web/database/index.ts
@@ -1,0 +1,1 @@
+export { DatabaseModule } from './database.module'
